Refactor for wine_categories to have show for wines and foods as nested controller.
