+++ package.json
@@ -14,6 +14,7 @@
     "knex": "^0.14.3",
     "morgan": "~1.9.0",
     "pg": "^7.4.1",
+    "pryjs": "^1.0.3",
     "serve-favicon": "~2.4.5"
   }
 }
