+++ routes/api/v1/wine_categories.js
@@ -3,9 +3,9 @@ const express = require('express');
 const router = express.Router();
 const configuration = require('../../../knexfile')[environment]
 const wineCategoriesController = require('../../../controllers/wine_categories_controller')
-
+const wineCategoriesWine = require('../../../controllers/wine_categories/wines')
 
 router.get('/', wineCategoriesController.index)
-router.get('/:id/wines', wineCategoriesController.show)
+router.get('/:id/wines', wineCategoriesWine.show)
 
 module.exports = router
